Refactor and build release
