+++ .github/workflows/release.yml
@@ -1,0 +1,24 @@
+name: Build Cloud Function Release
+
+on:
+  push:
+    tags:
+      - 'v*.*.*'
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Run Build Script
+        run: ./build.sh
+
+      - name: Release
+        uses: softprops/action-gh-release@v1
+        with:
+          files: oapen-irus-uk-cloud-function.zip
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
