+++ requirements.txt
@@ -1,2 +1,4 @@
-httpretty==1.0.*
-click==7.1.*
+requests==2.24.*
+geoip2==4.1.*
+jsonlines==1.2.*
+google-cloud-storage==1.35.*
