+++ tests/requirements.txt
@@ -1,0 +1,5 @@
+liccheck==0.4.*
+flake8==3.8.*
+coverage==5.2
+httpretty==1.0.*
+click==7.1.*
